feat(search): cleans search ab test
