+++ package.json
@@ -67,7 +67,7 @@
   },
   "devDependencies": {
     "@biomejs/biome": "2.4.16",
-    "@commitlint/cli": "21.1.0",
+    "@commitlint/cli": "21.2.1",
     "@commitlint/config-conventional": "21.2.0",
     "@fullhuman/postcss-purgecss": "8.0.0",
     "@playwright/test": "1.61.1",
