+++ types.ts
@@ -5,7 +5,6 @@ export type Entries<T> = {
 export interface MatomoAbTestVariations {
   AgentHeader: "original" | "VariationA";
   AgentWall: "original" | "VariationA" | "VariationB";
-  FiltresAvances: "original" | "VariationA" | "VariationB";
 }
 
 declare global {
